refactor: Move classes to their own directories
