+++ runner/src/hasura-client/index.ts
@@ -1,0 +1,1 @@
+export { default } from './hasura-client';
